requirements.txt
--extra-index-url https://gitlab.inria.fr/api/v4/projects/18692/packages/pypi/simple
huggingface_hub==0.30.0
torch==2.5.1
torchvision==0.20.1
torchdiffeq==0.2.5
accelerate==0.30.1
diffusers==0.26.3
transformers==4.53.3
einops==0.8.1
safetensors==0.5.3
bitsandbytes==0.49.0
numpy>=1.24.0,<2.0
scipy>=1.10.0
transforms3d==0.4.2
PyYAML==6.0
omegaconf==2.3.0
click==8.1.3
requests==2.32.4
openai==1.78.1
fbxsdkpy==2020.1.post2
trimesh==4.5.3
pyrender==0.1.45
pillow==11.1.0
gradio==6.0
